requirements.txt
mysql-connector-python==8.0.32
reportlab==4.0.4
Pillow==10.0.0
customtkinter==5.2.0
ttkbootstrap==1.10.1 